+++ lib/utils/UI.dart
@@ -127,6 +127,21 @@ class AppUI {
                   try {
                     String url = versions['android']['url'];
                     UpdateApp.updateApp(url: url, appleId: "1520301768");
+                    showCupertinoDialog(
+                        context: context,
+                        builder: (BuildContext ctx) {
+                          return CupertinoAlertDialog(
+                            title: Text(dic['update.download']),
+                            content: Text(dic['update.download.check']),
+                            actions: [
+                              CupertinoButton(
+                                child: Text(I18n.of(context)
+                                    .getDic(i18n_full_dic_ui, 'common')['ok']),
+                                onPressed: () => Navigator.of(ctx).pop(),
+                              ),
+                            ],
+                          );
+                        });
                   } catch (e) {
                     print('Failed to make OTA update. Details: $e');
                   }